Crate noirc_abi

Modules

errors

input_parser

serialization 🔒

Structs

Abi

AbiParameter

An argument or return value of the circuit’s main function.

AbiReturnType

Enums

AbiErrorType

AbiType

Types that are allowed in the (main function in binary)

AbiValue

AbiVisibility

Represents whether the parameter is public or known only to the prover.

Sign

Constants

MAIN_RETURN_NAME

Functions

decode_value

display_abi_error

Decode the raw revert data according to some error type descriptor.

Type Aliases

InputMap

A map from the fields in an TOML/JSON file which correspond to some ABI to their values