noirc_abi

Module input_parser

Modules

json

toml 🔒

Enums

Format

The different formats that are supported when parsing the initial witness values

InputTypecheckingError

InputValue

This is what all formats eventually transform into For example, a toml file will parse into TomlTypes and those TomlTypes will be mapped to Value

Functions

field_from_big_int 🔒

field_from_big_uint 🔒

field_to_signed_hex 🔒

parse_integer_to_signed 🔒

parse_str_to_field 🔒

parse_str_to_signed 🔒